Move and re-namespace Parser tests
