+++ EconDataLens.Tests/EtlParserTests/CpiDataParserTests.cs
@@ -5,7 +5,7 @@ using EconDataLens.Services;
 using Microsoft.Extensions.Options;
 
 
-namespace EconDataLens.Tests.ServiceTests;
+namespace EconDataLens.Tests.EtlParserTests;
 
 public class CpiDataParserTests
 {
