+++ EconDataLens.Tests/EtlParserTests/CpiItemParserTests.cs
@@ -4,7 +4,7 @@ using EconDataLens.Core.Configuration;
 using EconDataLens.Services;
 using Microsoft.Extensions.Options;
 
-namespace EconDataLens.Tests.ServiceTests;
+namespace EconDataLens.Tests.EtlParserTests;
 
 public class CpiItemParserTests
 {
